+++ src/x-request/index.ts
@@ -1,12 +1,26 @@
 import XFetch from './x-fetch';
-import XRequest from './x-request';
+import XRequest, {
+  getXRequestGlobalOptions,
+  resetXRequestGlobalOptions,
+  setXRequestGlobalOptions,
+} from './x-request';
 
 export type { XFetchMiddlewares, XFetchOptions, XFetchType } from './x-fetch';
-export type { XRequestBaseOptions, XRequestOptions, XRequestParams, XRequestCallbacks, XRequestFunction } from './x-request';
+export type {
+  XRequestBaseOptions,
+  XRequestOptions,
+  XRequestGlobalOptions,
+  XRequestParams,
+  XRequestCallbacks,
+  XRequestFunction,
+} from './x-request';
 
 export {
   XFetch,
-  XRequest
-}
+  XRequest,
+  setXRequestGlobalOptions,
+  getXRequestGlobalOptions,
+  resetXRequestGlobalOptions,
+};
 
 export default XRequest;